requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openpyxl==3.1.5
sentence-transformers==3.3.1
faiss-gpu-cu12
openai==1.57.4
numpy==1.26.4
pydantic==2.10.3
python-dotenv==1.0.1
aiosqlite==0.20.0
